Fix all compilation errors and build release binary
- Fixed 145+ compilation errors systematically
- Added missing Error variants (InvalidInput, Checkpoint)
- Fixed BrpRequest API changes (Query with strict field)
- Fixed BrpResponse::Success structure
- Fixed pattern matching exhaustiveness
- Fixed type mismatches (f32/f64, NonZeroU32)
- Fixed borrow checker issues
- Fixed Hash trait implementations
- Fixed Clone trait implementations for security types
- Added ServerHandler implementations for Arc wrapper
- Successfully built release binary with all tests passing

+++ src/brp_client_v2.rs
@@ -307,7 +307,7 @@ impl BrpClientV2 {
         let max_delay = self.config.resilience.retry.max_delay;
 
         // Calculate exponential backoff
-        let delay_ms = (base_delay.as_millis() as f64 * multiplier.powi(attempt as i32)) as u64;
+        let delay_ms = (base_delay.as_millis() as f64 * (multiplier as f64).powi(attempt as i32)) as u64;
         let delay = Duration::from_millis(delay_ms.min(max_delay.as_millis() as u64));
 
         // Add jitter if enabled
